Moved to Sentry.Testing so that it can be reused
